Complete change to what was planned
Originally the game was planned to be played interactively however, my
evil brain thought of something new. Why not just let it play itself!
Yep that's what's changed, now all you have to do is enter the name
of the players in arguments and the game will play on it's own and
declare the winner. Fancy your luck?